update: move resource validation to client service
