--- Authorization/Infrastructure/Services/Abstract/IResourceService.cs
@@ -1,8 +1,0 @@
-﻿using Application.Validation;
-
-namespace Infrastructure.Services.Abstract;
-
-public interface IResourceService
-{
-    Task<BaseValidationResult> ValidateResources(ICollection<string> resource, string scope);
-}
